Fixed compilation issues. Renamed all USE_XXX with the new HAVE_XXX generated by autoconf.

+++ hdt-lib/src/dictionary/KyotoDictionary.cpp
@@ -336,7 +336,7 @@ IteratorUCharString *KyotoDictionary::getShared() {
 	return new KyotoDictIterator(&this->shared);
 }
 
-unsigned int KyotoDictionary::getNumberOfElements()
+size_t KyotoDictionary::getNumberOfElements()
 {
 	return subjects.count()+predicates.count()+objects.count()+shared.count();
 }
